Decouple the Execlet abstraction from the type of the future.
What this means is that, inside a C++ coroutine, you can `co_await` a Folly
semifuture without bridging the return type of that semifuture to Rust.
Previously, you could only `co_await` a semifuture of the same type as the
coroutine return type, which was a very annoying limitation.

Happily, this ends up being an overall logic simplification.