Fix handling of exceptions in Streams.
Exceptions are delivered *synchronously* in C++ coroutines (via the
`unhandled_exception()` method), so we have to always reserve a dedicated slot
for them in our SPSC channel. Otherwise, in the case of a stream, the channel
could be full when we try to throw and we would have no way to wake up. This
commit implements this logic.