Change the macro to require the `unsafe` keyword.
The `cxx` project wants `unsafe` to be used whenever it's possible to have a
mismatch of types between C++ and Rust. Since there's no reasonable way for us
to check that compatible output types of each future or stream have been
supplied on either side of the bridge at present, this applies to us.