Fix a race in Folly execlet reference count manipulation that could cause memory leaks.
The attempt to have the Rust-side reference count account for both the C++-side
and Rust-side references was too clever by half. The problem is that there are
two allocations involved--one on the C++ side and one on the Rust side--and if
the Rust side happened to be destroyed first it had no way of knowing about the
one on the C++ side. This could cause memory leaks, which were readily
reproducible via Leak Sanitizer.